seperate points to it's own file
