+++ src/CMakeLists.txt
@@ -99,6 +99,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/verify.c"
         "${CMAKE_CURRENT_LIST_DIR}/serialized-utilities.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/frame.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/points.cpp"
 
         "${CMAKE_CURRENT_LIST_DIR}/algo.h"
         "${CMAKE_CURRENT_LIST_DIR}/api.h"
@@ -138,6 +139,7 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/serialized-utilities.h"
         "${CMAKE_CURRENT_LIST_DIR}/frame.h"
         "${CMAKE_CURRENT_LIST_DIR}/composite-frame.h"
+        "${CMAKE_CURRENT_LIST_DIR}/points.h"
         "${CMAKE_CURRENT_LIST_DIR}/depth-sensor.h"
         "${CMAKE_CURRENT_LIST_DIR}/color-sensor.h"
 )
