+++ src/frame.h
@@ -262,18 +262,6 @@ private:
     std::shared_ptr< stream_profile_interface > stream;
 };
 
-class points : public frame
-{
-public:
-    float3 * get_vertices();
-    void export_to_ply( const std::string & fname, const frame_holder & texture );
-    size_t get_vertex_count() const;
-    float2 * get_texture_coordinates();
-};
-
-MAP_EXTENSION( RS2_EXTENSION_POINTS, librealsense::points );
-
-
 class video_frame : public frame
 {
 public:
